+++ RELEASE-NOTES.md
@@ -23,6 +23,7 @@
 * Added connection option `--token-file-path` allowing passing OAuth token using a file. The function is also
   supported by setting `token_file_path` in connection definition.
 * Support for Python remote execution via `snow stage execute` and `snow git execute` similar to existing EXECUTE IMMEDIATE support.
+* Added support for autocomplete in `--connection` flag.
 
 ## Fixes and improvements
 * The `snow app run` command now allows upgrading to unversioned mode from a versioned or release mode application installation
@@ -31,6 +32,7 @@
 * Added support for user stages in stage execute command
 * Added support for user stages in stage and git copy commands
 * Improved support for quoted identifiers in snowpark commands.
+
 
 # v2.6.0
 ## Backward incompatibility